Split DMRG-specific code into its own file
- Now solver.jl is entirely abstract
